+++ src/components/IsPrivate.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { AuthContext } from "../context/auth.context";
+import { AuthContext } from "../context/AuthContext";
 import { Navigate } from "react-router-dom";
 import Spinner from "react-bootstrap/Spinner";
